+++ day05-1/src/conversion.rs
@@ -1,12 +1,12 @@
 #[derive(Debug)]
 pub struct Conversion {
-    pub destination_start: u32,
-    pub source_start: u32,
-    pub range_length: u32
+    pub destination_start: u64,
+    pub source_start: u64,
+    pub range_length: u64
 }
 
 impl Conversion {
-    pub fn new(destination_start: u32, source_start: u32, range_length: u32) -> Conversion {
+    pub fn new(destination_start: u64, source_start: u64, range_length: u64) -> Conversion {
         Conversion {
             destination_start: destination_start,
             source_start: source_start,
@@ -14,7 +14,7 @@ impl Conversion {
         }
     }
 
-    pub fn convert(&self, input: &u32) -> Option<u32> {
+    pub fn convert(&self, input: &u64) -> Option<u64> {
         let range = self.source_start..(self.source_start + self.range_length);
         if range.contains(&input) {
             let position = input - self.source_start;
